+++ app/src/main/java/com/mapbox/maps/testapp/examples/CircleLayerClusteringActivity.kt
@@ -34,35 +34,35 @@ class CircleLayerClusteringActivity : AppCompatActivity() {
     val mapboxMap = mapView.getMapboxMap()
 
     mapboxMap.loadStyleUri(
-      Style.LIGHT
-    ) {
+      styleUri = Style.LIGHT,
       // Disable any type of fading transition when icons collide on the map. This enhances the visual
       // look of the data clustering together and breaking apart.
-      it.styleTransition = TransitionOptions.Builder()
+      styleTransitionOptions = TransitionOptions.Builder()
         .duration(0)
         .delay(0)
         .enablePlacementTransitions(false)
-        .build()
+        .build(),
+      onStyleLoaded = {
+        mapboxMap.flyTo(
+          CameraOptions.Builder()
+            .center(Point.fromLngLat(-79.045, 12.099))
+            .zoom(3.0)
+            .build()
+        )
 
-      mapboxMap.flyTo(
-        CameraOptions.Builder()
-          .center(Point.fromLngLat(-79.045, 12.099))
-          .zoom(3.0)
-          .build()
-      )
+        addClusteredGeoJsonSource(it)
 
-      addClusteredGeoJsonSource(it)
+        bitmapFromDrawableRes(this, R.drawable.ic_cross)?.let { bitmap ->
+          it.addImage(CROSS_ICON_ID, bitmap, true)
+        }
 
-      bitmapFromDrawableRes(this, R.drawable.ic_cross)?.let { bitmap ->
-        it.addImage(CROSS_ICON_ID, bitmap, true)
+        Toast.makeText(
+          this@CircleLayerClusteringActivity,
+          R.string.zoom_map_in_and_out_instruction,
+          Toast.LENGTH_SHORT
+        ).show()
       }
-
-      Toast.makeText(
-        this@CircleLayerClusteringActivity,
-        R.string.zoom_map_in_and_out_instruction,
-        Toast.LENGTH_SHORT
-      ).show()
-    }
+    )
   }
 
   private fun addClusteredGeoJsonSource(style: Style) {