+++ sdk/src/test/java/com/mapbox/maps/MapboxMapTest.kt
@@ -5,6 +5,7 @@ import com.mapbox.bindgen.Value
 import com.mapbox.common.ShadowLogger
 import com.mapbox.geojson.Feature
 import com.mapbox.geojson.Point
+import com.mapbox.maps.extension.observable.eventdata.MapLoadingErrorEventData
 import com.mapbox.maps.extension.style.StyleContract
 import com.mapbox.maps.extension.style.style
 import com.mapbox.maps.plugin.MapProjection
@@ -34,6 +35,7 @@ class MapboxMapTest {
   private val nativeObserver: NativeObserver = mockk(relaxed = true)
   private val resourceOptions = mockk<ResourceOptions>(relaxed = true)
 
+  private lateinit var styleObserver: StyleObserver
   private lateinit var mapboxMap: MapboxMap
 
   @Before
@@ -41,7 +43,8 @@ class MapboxMapTest {
     mockkStatic(Map::class)
     every { Map.clearData(any(), any()) } just runs
     every { nativeMap.resourceOptions } returns resourceOptions
-    mapboxMap = MapboxMap(WeakReference(nativeMap), nativeObserver, 1.0f)
+    styleObserver = mockk(relaxUnitFun = true)
+    mapboxMap = MapboxMap(WeakReference(nativeMap), nativeObserver, styleObserver)
   }
 
   @Test
@@ -1032,5 +1035,21 @@ class MapboxMapTest {
     val memoryBudget = mockk<MapMemoryBudget>()
     mapboxMap.setMemoryBudget(memoryBudget)
     verify { nativeMap.setMemoryBudget(memoryBudget) }
+  }
+
+  @Test
+  fun setStyleTransition() {
+    val options = mockk<TransitionOptions>(relaxed = true)
+    mapboxMap.loadStyleUri(
+      "style",
+      options,
+      {},
+      object : OnMapLoadErrorListener {
+        override fun onMapLoadError(eventData: MapLoadingErrorEventData) {}
+      }
+    )
+    verify(exactly = 1) {
+      styleObserver.setLoadStyleListener(options, any(), any())
+    }
   }
 }